move spectrum/filter convolving to new file
